+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### Added
+- `.clone` method now exists to create a clone of the Config object.
+
 ## [2.0.0] - 2018-08-26
 ### Changed
 - `.read` will now return the `Config` object, instead of a `Bool`.